Add a prompt to collect this week's GH activity
It's nice to be able to fire off a request to collect my GitHub
activity so I can come back to it and easily see what I've
contributed to over the past week. I can use this to remember anything I
may want to demo at the end of the week or to remember to add something
to a brag doc.